+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+
+## 2.3.0 - 2022-04-05
 - Add `ImpureResponseDecoderInterface`
 - Show detailed type for `DecodedData` value object